Refactor - Update NetworkMultiSearchListItem's voteAverage parameter to be a double.

+++ app/src/main/java/com/github/androidpirate/capsulereviews/data/network/response/multiSearch/NetworkMultiSearchListItem.kt
@@ -27,5 +27,5 @@ data class NetworkMultiSearchListItem(
     @SerializedName("video")
     val video: Boolean,
     @SerializedName("vote_average")
-    val voteAverage: Int
+    val voteAverage: Double
 )
